Refactor project structure for tests and page classes
