+++ src/test/java/site/GatewayTest.java
@@ -1,4 +1,4 @@
-package site.alpha;
+package site;
 
 import org.junit.After;
 import org.junit.Before;
@@ -7,9 +7,8 @@ import org.junit.Test;
 import ru.yandex.qatools.allure.annotations.Description;
 import ru.yandex.qatools.allure.annotations.Features;
 import ru.yandex.qatools.allure.annotations.Stories;
-import site.BaseTest;
-import site.alpha.pages.GatewayPage;
-import site.alpha.pages.authentication.LoginPage;
+import site.pages.GatewayPage;
+import site.pages.authentication.LoginPage;
 
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.core.Is.is;
