+++ src/test/java/site/authentication/LoginTest.java
@@ -1,4 +1,4 @@
-package site.alpha;
+package site.authentication;
 
 import org.junit.After;
 import org.junit.Before;
@@ -6,8 +6,9 @@ import org.junit.Test;
 import ru.yandex.qatools.allure.annotations.Description;
 import ru.yandex.qatools.allure.annotations.Features;
 import ru.yandex.qatools.allure.annotations.Stories;
+import site.Accounts;
 import site.BaseTest;
-import site.alpha.pages.authentication.LoginPage;
+import site.pages.authentication.LoginPage;
 import utilities.users.User;
 
 import static org.hamcrest.MatcherAssert.assertThat;
